Alcohol Consumption per Capita

These data provide per capita alcohol consumption values for many countries in 2005 and 2008. There are also a few countries for which there are data in other years.

data(Alcohol)

Format

A data frame with 411 observations on the following variables.

country country name

year year

alcohol estimated per capita alcohol consumption for adults (15+) in litres pure alcohol

Source

Gapminder (https://www.gapminder.org/data/)

Examples

data(Alcohol)
# There are only a few observations in years other than 2005 and 2008
subset(Alcohol, ! year %in% c(2005,2008))

#>               country year alcohol
#> 1              Russia 1985   13.31
#> 2              Russia 1986   10.77
#> 3              Russia 1987   10.96
#> 4              Russia 1988   11.57
#> 5              Russia 1989   12.04
#> 6               Chile 1990    9.43
#> 7             Ecuador 1990    8.40
#> 8              Greece 1990   12.51
#> 9              Russia 1990   12.29
#> 10             Russia 1991   12.67
#> 11  Macedonia [FYROM] 1992   12.66
#> 12             Russia 1992   13.23
#> 13            Moldova 1993   18.10
#> 14             Russia 1993   13.60
#> 15           Slovenia 1993   24.19
#> 16            Ukraine 1993   13.00
#> 17            Austria 1994   12.62
#> 18            Denmark 1994   14.36
#> 19            Finland 1994    9.79
#> 20            Iceland 1994    5.52
#> 21             Norway 1994    6.26
#> 22             Russia 1994   13.30
#> 23            Estonia 1995   10.74
#> 24            Hungary 1995   14.52
#> 25        New Zealand 1995    9.19
#> 26       South Africa 1995   10.00
#> 27             Brazil 1996   14.01
#> 28             Sweden 1996    8.00
#> 29             Sweden 1998    8.20
#> 30             Sweden 2000    8.40
#> 31             Sweden 2001    9.10
#> 32             Sweden 2002    9.90
#> 33             Sweden 2003   10.20
#> 34             Sweden 2004   10.50
#> 223            Sweden 2006    9.80
#> 224            Sweden 2007    9.80