Foot measurements in children

These data were collected by a statistician, Mary C. Meyer, in a fourth grade classroom in Ann Arbor, MI, in October 1997. They are a convenience sample --- the kids who were in the fourth grade.

data(KidsFeet)

Format

A data frame with 39 observations on the following variables.

name

a factor with levels corresponding to the name of each child

birthmonth

the month of birth

birthyear

the year of birth

length

length of longer foot (in cm)

width

width of longer foot (in cm)

sex

a factor with levels B G

biggerfoot

a factor with levels L R

domhand

a factor with levels L R

Details

Quoted from the source: "From a very young age, shoes for boys tend to be wider than shoes for girls. Is this because boys have wider feet, or because it is assumed that girls, even in elementary school, are willing to sacrifice comfort for fashion? To assess the former, a statistician measures kids' feet."

References

Mary C. Meyer (2006) "Wider Shoes for Wider Feet?" Journal of Statistics Education 14(1), http://jse.amstat.org/v14n1/datasets.meyer.html.

Examples

data(KidsFeet)