Volume of Users of a Massachusetts Rail Trail

The Pioneer Valley Planning Commission (PVPC) collected data north of Chestnut Street in Florence, MA for ninety days from April 5, 2005 to November 15, 2005. Data collectors set up a laser sensor, with breaks in the laser beam recording when a rail-trail user passed the data collection station.

data(Riders)

Format

A data frame with 90 observations on the following 12 variables.

date

date of data collection (POSIXct)

day

a factor with levels Monday, Tuesday, Wednesday, Thursday, Friday, Saturday, and Sunday.

highT

high temperature for the day (in degrees Fahrenheit)

lowT

low temperature for the day (in degrees Fahrenheit)

hi

shorter name for highT

lo

shorter name for lowT

precip

inches of precipitation

clouds

measure of cloud cover (in oktas)

riders

estimated number of trail crossings that day (number of breaks recorded)

ct

shorter name for riders

weekday

type of day: a factor with levels N (weekend or holiday) Y (non-holiday weekday)

wday

shorter name for weekday

Source

Pioneer Valley Planning Commission, http://www.fvgreenway.org/pdfs/Northampton-Bikepath-Volume-Counts%20_05_LTA.pdf

Details

There is a potential for error when two users trigger the infrared beam at exactly the same time since the counter would only logs one of the crossings. The collectors left the motion detector out during the winter, but because the counter drops data when the temperature falls below 14 degrees Fahrenheit, there are no data for the coldest winter months.

References

"Rail trails and property values: Is there an association?", Nicholas J. Horton and Ella Hartenian (Journal of Statistics Education, 2015), http://www.amstat.org/publications/jse/v23n2/horton.pdf

Examples

data(Riders)
str(Riders)

#> 'data.frame':	90 obs. of  12 variables:
#>  $ date   : POSIXct, format: "2005-04-05" "2005-04-06" ...
#>  $ day    : Factor w/ 7 levels "Friday","Monday",..: 6 7 5 1 3 4 2 6 7 5 ...
#>  $ highT  : int  62 75 70 65 66 74 56 54 59 64 ...
#>  $ lowT   : int  39 43 47 48 38 33 38 32 34 41 ...
#>  $ hi     : int  62 75 70 65 66 74 56 54 59 64 ...
#>  $ lo     : int  39 43 47 48 38 33 38 32 34 41 ...
#>  $ precip : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ clouds : num  0.8 6.3 9.9 7.2 0 2.5 0 3.6 2.4 3.4 ...
#>  $ riders : int  236 156 328 418 629 635 335 304 344 349 ...
#>  $ ct     : int  236 156 328 418 629 635 335 304 344 349 ...
#>  $ weekday: Factor w/ 2 levels "N","Y": 2 2 2 2 1 1 2 2 2 2 ...
#>  $ wday   : Factor w/ 2 levels "N","Y": 2 2 2 2 1 1 2 2 2 2 ...