Countries

A data frame containing country names as used by Gapminder and the maps package to facilitate conversation between the two.

data(Countries)

Format

A data frame with 258 observations on the following variables.

maptools

region name http://mappinghacks.com/ data sets

gapminder

country name in Gapminder data sets

maps

region name in maps data sets

Details

The "countries" in the maps data include several other geographic regions (bodies of water, islands belonging to other countries, Hawaii, etc.) that are not countries. Furthermore, the maps countries do not include many of the countries that have been created since ca. 2000. The mapping is therefore many-to-many, and also includes some NAs when there is no appropriate mapping. Bodies of water in the maps data, for example, are not assigned a country in the Gapminder.

Examples

data(Countries)
subset(Countries, maps=="Yugoslavia")  # Where has Yugoslavia gone?
#>                   maptools       maps              gapminder
#> 33  Bosnia and Herzegovina Yugoslavia Bosnia and Herzegovina
#> 64                 Croatia Yugoslavia                Croatia
#> 227                 Serbia Yugoslavia                 Serbia
#> 234               Slovenia Yugoslavia               Slovenia
subset(Countries, is.na(gapminder))    # Things from maps with no Gapminder equivalent
#>                                 maptools                     maps gapminder
#> 2                          Åland Islands                     <NA>      <NA>
#> 6                                   <NA>          Andaman Islands      <NA>
#> 9                               Anguilla                 Anguilla      <NA>
#> 10                            Antarctica               Antarctica      <NA>
#> 12                                  <NA>                 Aral Sea      <NA>
#> 15                                 Aruba                    Aruba      <NA>
#> 19                                  <NA>                   Azores      <NA>
#> 30                                  <NA>                Black Sea      <NA>
#> 32                                  <NA>                  Bonaire      <NA>
#> 35                         Bouvet Island                     <NA>      <NA>
#> 37        British Indian Ocean Territory                     <NA>      <NA>
#> 41                               Bermuda                     <NA>      <NA>
#> 47                                  <NA>           Canary Islands      <NA>
#> 49                                  <NA>              Caspian Sea      <NA>
#> 50                        Cayman Islands           Cayman Islands      <NA>
#> 53                                  <NA>       Chagos Archipelago      <NA>
#> 56                      Christmas Island                     <NA>      <NA>
#> 57               Cocos (Keeling) Islands                     <NA>      <NA>
#> 66                                  <NA>                  Curacao      <NA>
#> 80           Falkland Islands (Malvinas)         Falkland Islands      <NA>
#> 81                         Faroe Islands                     <NA>      <NA>
#> 85                         French Guiana            French Guiana      <NA>
#> 86                      French Polynesia         French Polynesia      <NA>
#> 87   French Southern and Antarctic Lands                     <NA>      <NA>
#> 90                                  <NA>               Gaza Strip      <NA>
#> 92  South Georgia South Sandwich Islands                     <NA>      <NA>
#> 95                             Gibraltar                     <NA>      <NA>
#> 96                                  <NA>        Great Bitter Lake      <NA>
#> 97                                  <NA>              Great Lakes      <NA>
#> 99                             Greenland                Greenland      <NA>
#> 101                           Guadeloupe               Guadeloupe      <NA>
#> 102                                 Guam                     <NA>      <NA>
#> 109    Heard Island and McDonald Islands                     <NA>      <NA>
#> 111              Holy See (Vatican City)                     <NA>      <NA>
#> 112                            Hong Kong                     <NA>      <NA>
#> 120                                 <NA>               Irian Jaya      <NA>
#> 121                          Isle of Man              Isle of Man      <NA>
#> 122                                 <NA>            Isle of Wight      <NA>
#> 134                                 <NA>             Lacul Greaca      <NA>
#> 135                                 <NA>              Lake Albert      <NA>
#> 136                                 <NA>         Lake Fjerritslev      <NA>
#> 137                                 <NA>              Lake Kariba      <NA>
#> 138                                 <NA>              Lake Malawi      <NA>
#> 139                                 <NA>           Lake Pasvikelv      <NA>
#> 140                                 <NA>          Lake Tanganyika      <NA>
#> 141                                 <NA>            Lake Titicaca      <NA>
#> 142                                 <NA>            Lake Victoria      <NA>
#> 149                        Liechtenstein            Liechtenstein      <NA>
#> 153                                Macau                     <NA>      <NA>
#> 155                                 <NA>          Madeira Islands      <NA>
#> 158                             Maldives                 Maldives      <NA>
#> 162                     Marshall Islands         Marshall Islands      <NA>
#> 163                           Martinique               Martinique      <NA>
#> 164                                 <NA>              Maug Island      <NA>
#> 167                              Mayotte                     <NA>      <NA>
#> 171                               Monaco                   Monaco      <NA>
#> 173                           Montenegro                     <NA>      <NA>
#> 174                           Montserrat               Montserrat      <NA>
#> 182                 Netherlands Antilles                     <NA>      <NA>
#> 183                                 <NA>             Neutral Zone      <NA>
#> 184                                 <NA>                    Nevis      <NA>
#> 185                        New Caledonia            New Caledonia      <NA>
#> 193                       Norfolk Island                     <NA>      <NA>
#> 194             Northern Mariana Islands Northern Mariana Islands      <NA>
#> 200                            Palestine                     <NA>      <NA>
#> 202                                 <NA>          Paracel Islands      <NA>
#> 206                     Pitcairn Islands         Pitcairn Islands      <NA>
#> 211                              Reunion                     <NA>      <NA>
#> 214                                 <NA>          Saint Eustatius      <NA>
#> 215                         Saint Helena                     <NA>      <NA>
#> 219                     Saint Barthelemy         Saint-Barthelemy      <NA>
#> 220            Saint Pierre and Miquelon                     <NA>      <NA>
#> 221                         Saint Martin             Saint-Martin      <NA>
#> 224                                 <NA>                 Sardinia      <NA>
#> 229                                 <NA>                   Sicily      <NA>
#> 231                                 <NA>          Sin Cowe Island      <NA>
#> 237                                 <NA>          Sonsorol Island      <NA>
#> 241                                 <NA>   South Sandwich Islands      <NA>
#> 243                                 <NA>           Spratly Island      <NA>
#> 247                             Svalbard                     <NA>      <NA>
#> 252                               Taiwan                     <NA>      <NA>
#> 259                              Tokelau                  Tokelau      <NA>
#> 265             Turks and Caicos Islands         Turks and Caicos      <NA>
#> 273 United States Minor Outlying Islands                     <NA>      <NA>
#> 279               British Virgin Islands           Virgin Islands      <NA>
#> 280         United States Virgin Islands           Virgin Islands      <NA>
#> 281                                 <NA>         Vislinskiy Zaliv      <NA>
#> 283            Wallis and Futuna Islands                     <NA>      <NA>
#> 285                       Western Sahara           Western Sahara      <NA>
subset(Countries, is.na(maps))         # Things from Gapminder with no maps equivalent
#>                                 maptools maps    gapminder
#> 2                          Åland Islands <NA>         <NA>
#> 14                               Armenia <NA>      Armenia
#> 18                            Azerbaijan <NA>   Azerbaijan
#> 35                         Bouvet Island <NA>         <NA>
#> 37        British Indian Ocean Territory <NA>         <NA>
#> 41                               Bermuda <NA>         <NA>
#> 56                      Christmas Island <NA>         <NA>
#> 57               Cocos (Keeling) Islands <NA>         <NA>
#> 76                               Eritrea <NA>      Eritrea
#> 78                               Estonia <NA>      Estonia
#> 81                         Faroe Islands <NA>         <NA>
#> 87   French Southern and Antarctic Lands <NA>         <NA>
#> 91                               Georgia <NA>      Georgia
#> 92  South Georgia South Sandwich Islands <NA>         <NA>
#> 95                             Gibraltar <NA>         <NA>
#> 102                                 Guam <NA>         <NA>
#> 109    Heard Island and McDonald Islands <NA>         <NA>
#> 111              Holy See (Vatican City) <NA>         <NA>
#> 112                            Hong Kong <NA>         <NA>
#> 129                           Kazakhstan <NA>   Kazakhstan
#> 132                           Kyrgyzstan <NA>   Kyrgyzstan
#> 144                               Latvia <NA>       Latvia
#> 150                            Lithuania <NA>    Lithuania
#> 153                                Macau <NA>         <NA>
#> 159                  Republic of Moldova <NA>      Moldova
#> 167                              Mayotte <NA>         <NA>
#> 173                           Montenegro <NA>         <NA>
#> 182                 Netherlands Antilles <NA>         <NA>
#> 190                                 Niue <NA>         Niue
#> 193                       Norfolk Island <NA>         <NA>
#> 199                                Palau <NA>        Palau
#> 200                            Palestine <NA>         <NA>
#> 211                              Reunion <NA>         <NA>
#> 215                         Saint Helena <NA>         <NA>
#> 220            Saint Pierre and Miquelon <NA>         <NA>
#> 232                            Singapore <NA>    Singapore
#> 247                             Svalbard <NA>         <NA>
#> 252                               Taiwan <NA>         <NA>
#> 253                           Tajikistan <NA>   Tajikistan
#> 256                          Timor-Leste <NA>  Timor-Leste
#> 264                         Turkmenistan <NA> Turkmenistan
#> 269                              Ukraine <NA>      Ukraine
#> 273 United States Minor Outlying Islands <NA>         <NA>
#> 274                           Uzbekistan <NA>   Uzbekistan
#> 283            Wallis and Futuna Islands <NA>         <NA>