Houses in Saratoga County (2006)

Data on houses in Saratoga County, New York, USA in 2006

data(SaratogaHouses)

Format

A data frame with 1728 observations on the following 16 variables.

price price (US dollars)

lotSize size of lot (acres)

age age of house (years)

landValue value of land (US dollars)

livingArea living are (square feet)

pctCollege percent of neighborhood that graduated college

bedrooms number of bedrooms

firplaces number of fireplaces

bathrooms number of bathrooms (half bathrooms have no shower or tub)

rooms number of rooms

heating type of heating system

fuel fuel used for heating

sewer type of sewer system

waterfront whether property includes waterfront

newConstruction whether the property is a new construction

centralAir whether the house has central air

Source

Data collected by Candice Corvetti and used in the "Stat 101" case study "How much is a Fireplace Worth". See also https://www.saratogacountyny.gov/departments/real-property-tax-service-agency/